refactor(tests): break tests into different files add more test cases
