+++ server/shared/validateSignUpInput.js
@@ -10,19 +10,19 @@ const validateSignUpInput = (userInput) => {
   if (!validator.isMobilePhone(userInput.phoneNumber, 'any')) {
     errors.phoneNumber = 'Phone number not valid';
   }
-  if (!validator.matches(userInput.username, /^[a-zA-Z0-9_]*$/)) {
+  if (!validator.matches(userInput.username.trim(), /^[a-zA-Z0-9_]*$/)) {
     errors.username = 'Username cannot contain special characters aside from _';
   }
-  if (!validator.isLength(userInput.username, { min: 3 })) {
+  if (!validator.isLength(userInput.username.trim(), { min: 3 })) {
     errors.username = 'Username is too short';
   }
-  if (!validator.isLength(userInput.password, { min: 6 })) {
+  if (!validator.isLength(userInput.password.trim(), { min: 6 })) {
     errors.password = 'Password cannot be less than 6 characters';
   }
-  if (!validator.matches(userInput.fullName, /^[a-zA-Z ]*$/)) {
+  if (!validator.matches(userInput.fullName.trim(), /^[a-zA-Z ]*$/)) {
     errors.fullName = 'Name can only be alphabets';
   }
-  if (!validator.isLength(userInput.fullName, { min: 2 })) {
+  if (!validator.isLength(userInput.fullName.trim(), { min: 2 })) {
     errors.fullName = 'Name is too short';
   }
   if (!validator.isEmail(userInput.email)) {
